requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.1
-e git+https://github.com/patilhrishikesh/mlproject.git@e4f0cf326ece3b0dc6d4557f71e2491b2b849bf0#egg=mlproject
numpy==2.2.3
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.1
